Create new file for name trimmer class
