+++ interface.rb
@@ -3,20 +3,3 @@ class Nameable
     raise NotImplementedError
   end
 end
-
-
-class CapitalizeDecorator < BaseDecorator
-  def correct_name
-    @nameable.correct_name.capitalize
-  end
-end
-
-class TrimmerDecorator < BaseDecorator
-  def correct_name
-    if @nameable.correct_name.length > 10
-      @nameable.correct_name[0, 10]
-    else
-      @nameable.correct_name
-    end
-  end
-end
